+++ src/@shared/database/paginated-result.ts
