+++ src/@shared/http-client/http-client.interface.ts
